+++ src/main/java/org/cojen/maker/TheFieldMaker.java
@@ -37,6 +37,11 @@ final class TheFieldMaker extends ClassMember implements FieldMaker, Typed {
     }
 
     @Override
+    public String name() {
+        return mField.name();
+    }
+
+    @Override
     public FieldMaker public_() {
         mModifiers = Modifiers.toPublic(mModifiers);
         return this;